Get rid of ".ast" files in parser tests
Now the tests for the parser only look at the part of the AST related to
the feature they are testing and don't care if the structure of an
unrelated part changes.

This should hopefuly reduce the number of tests that break when we
update the AST, as well as the clutter in the patch diffs.