Temporarily stop using parser-gen
We were using parger-gen in the parser and plaing lpeglabel in the lexer
but unfortuantely they have incompatible APIs when it comes to error
labels (which are crucial for implementing syntax errors).

This commit converts the parser to plain lpeglabel. Hopefully we can go
back to parser-gen later when this issue is resolved in those libraries.