Improvements to parser_spec
- Inline the test cases instead of putting them on separate files
- AST tests for expressions / statements instead of full program
- Better error message when we get an unexpected syntax error
- Better error message when AST doesn't match what was expected

The original reason for putting the test cases in a separate file was
because assert.are.same wouldn't give a helpful error message when the
AST discrepancy was deep inside the AST structure so we resorted to
running titanc with the --print-ast option to debug those. But by
calling assert.are.same closer to the place that matters in the AST
structure we can get good error messages from it, se don't need to keep
the files separate anymore. Another thing that we gain by focusing the
tests on only the part of the code that matters is that test cases about
expressions and statements won't break if we change the syntax for
toplevel declarations.