Remove float->int coercions from bitwise operators
Although this is a coercion that Lua performs, I think it is better for
Titan to produce a compilation error in this case. float to int
coercions can fail with a runtime error so it is in the interest of the
programmer that Titan reject this by default and demand an explicit type
conversion if that is the true intent of the programmer.