Remove module tests from checker and from coder
This should allow us to temporarily remove modules from Titan, as part
of the plan to cut down complexity on this version of this prototype.
The most pressing need at the moment was that modules overload the "."
syntax to mean both run-time field access and compile-time namespaces.

I left the AST structure, the parser and the parser tests intact because
there is nothing to gain right now from removing that code and adding
that back later would be tricky.