Create a separate pass for variable scope analysis
Trying to do scope analysis and typechecking at the same time tried to
push the code in conflicting directions. For example, consider something
liks this

    local x = 1
    local y = 2
    local x: Point = { x = x, y = y }

In the third line, the scope analysis wants to bind the names in the
type annotation and initializer expression before adding the new x to
the scope. On the other hand, the typechecker wants to know the type of
x before typechecking expr so it can do top down type inference for the
initializer list.

This first commit creates a separate scope analysis pass and creates a
test suite for it but does not yet integrate with the rest of the
compiler. We still need to rewrite the typechecker to use the result of
this pass instead of duplicating it.