Get rid of luassert dependency in coder tests
The coder tests run in our patched version of Lua and being able to
require additional dependencies is tricky. This should let Travis run
the coder tests again.

The first time this bit us was in Travis, which sets the LUA_PATH to
ignore the system stuff and today I was bit again by this when I tried
to update our Lua to 5.4 (because my system Luassert is installed in a
5.3 path)