Remove tl_node.block trick from titan entry_point
This only served to remove a set of braces from the generated code, at the cost
of making the code hard to understand. Anothere sign that we shouldn't be
bothering with this is that the extra braces are going to go away anyway when we
switch the coder operate on an intermediate representation.