Put all luaL_error calls in separate functions
This changed the behavior of array writes. Now they check tags just like an
array read would.

There is no performance gain in allowing writes to slots containing invalid
types because we need to check the slot tag anyway to know if we are inside
or outside the array bounds.