Make intop and binop generate a var name
Initial commit in an ongoing effort to replace make sure that generate_exp
always returns a single varname or literal, which means that order of evaluation
is fully explicit and that the output of generate_exp can safely be used in
macros or macro-like things.

This patch also makes unary negation for integers have well-defined wraparound.