Make integer division generate a var name
This way the luaV_div function doesn't end up duplicated...

This patch also inlines the logic of luaV_div in the generated code, which can
result in a big performance boost if one of the parameters is a constant. This
was very noticeable in the binsearch benchmark (more than 2x improvement).