Checker spec now only cares about the first error
In preparation for removing support for multiple type error messages, now the
test suite for the type checker only looks at the first error message that is
generated.

I changed the few tests cases that produced multiple error messages and where
the error message that the test suite cared about was not the first one.