Desugar if-elseif into if-then-else
This commit gets rid of the awkward representation we have of if statments right
now, with the "Thens" array and optional else statement.

I believe the current representation started out as a side effect of the parser
algorithm and then we kept it made the generated code prettier:

```
/* old generated code */
if (c1) {
   ...
} else if (c2) {
   ...
}
```

instead of

```
/* obvious generated code*/
if (c1) {
} else {
    if (c2) {
    } else {
    }
}
```

However, this motivation doesn't apply anymore now that we are switching our
generated code to a "flat" list of basic blocks.