Change how benchmarks detect Luajit vs PUC Lua
Now the idea is that files with "luajit" in their name are Luajit-only and files
with "puc" in their name are only for PUC-Lua. Having a way to mark files that
need 5.3 features and therefore cannot run under LuaJIT sounds like a good idea.

This commit doesn't fully implement the "puc" part in the run_benchmarks
script, but it does work for the current set of benchmarks we have. I'm planning
on rewriting run_benchmarks in the very near future anyway so I didn't want to
worry too much about the current version of that code.