Split util.shell into two functions
util.execute runs a command, without capturing or redirecting its output. The
new name is meant to be a reference to os.execute, to make it clearer that this
function doesn't return the stdout as a string anymore.

util.outputs_of_execute is the new function when you need to use the stdout of
the command. It also obtains the result of stderr.