Refactor benchmarks script (single run only)
This commit replaces the current run_benchmarks script with a Lua
library and an executable script.

The executable script only runs a single benchmark once, and displays the
raw output of the measurement command (time or pef). It is meant for
interactive use.

The lua library is responsible for compiling the required executables
and actually running the benchmarks.

This commit does not yet implement scripts for running multiple
benchmarks. The plan is to, for each table in the paper, produce a
script that directly hooks up into the low-level library. This will
allow for fine-tuned control over what benchmarks to run, in what order
to run them, and the output format.