Revert to string _tags instead of function _tags
Closes #29. Now we do not need to call typedecl.tostring to be able to read what
the tag for a value is. The output of the "inspect" function is easier to read
now, which should help debugging. It also makes our code a tad simpler and less
magic.

The current rationale for function tags is twofold:

1. Avoid having to type "" in the if-then-else statements

2. Allow renaming the modules when requiring them. For example:

    ```lua
    local aaast = require 'pallene
    if blah._tag == aaast.Exp.Bool then
    ```

In hindsight, I think these are not very compelling reasons anymore. For (1), I
think the negative of needing to use typedecl.tostring outweighs the positive of
not needing to type quotes. For (2), the tostring function will still output
"ast.Exp.Bool" even if we reimport the ast module under a different name, so the
suggestion that we can rename it at will like that is actively misleading.

----

Another alternative solution that I explored was to replace function _tags by
special objects with a __call and __tostring metamethod. This would have been
less disruptive in terms of lines-of-code changed, but it would only add more
complexity to the code (instead of removing it) and it would also make the
"inspect" output clunkier, since it would now print those metatables. This
solution also would not fix the problem 2) that I mentioned above