Refactor the checker to use bidirectional style
This is a big restructuring in how we type check expressions. Instead of
having a single check_type function with an optional type_hint parameter, we
now have two functions: check_type_synthesize and check_type_verify.

check_type_synthesize is closer to our original version of check_type. It is
used when we want to compute the type of the expression "by itself".

check_type_verify is used in contexts where we have an expected result type.
This is the only way to typecheck Initlist nodes, and is used to automatically
insert coercions to and from value throughout our program.

---

The biggest impact of this change is that the functionality of check_match and
try_coerce were merged into the check_type_verify function. This means that
everywhere that we used to use check_match the new version of the typechecker
may now introduce a type coercions, which it wouldn't be before. This makes
our coercions behavior more consistent.

Other than that, the code also became somewhat easier to understand, as the
interface of the check_exp_ is much easier to reason about now, and we were
able to delete many of the helper functions like check_match and try_coerce.