Merge scope_analysis and checker
This is the first step in the big IR rewrite. Scope analysis and typechecking
are merged in a single pass that returns a Module node that organizes all the
types and functions found in the toplevel. (See ir.lua for details)

This commit passes the tests for parser_spec and checker_spec but everything
else is obviously broken. Test coverage on these is not complete so there are
likely to still be bugs to be ironed out.

_decl annotations have been replaced by _name annotations of type checker.Name.
This richer data type should make the code easier to understand, and provides
a better path towards implementing closures than the current approach of having
a single _decl node. (This would be done by adding a field to checker.Name.Local
with a pointer to the parent function).

_type annotations are still present, but only on ast.Var and ast.Exp nodes.
The annotations we used to have on toplevel nodes are now part of the ir.Module
and ir.Function records. (This also includes the _field_types annotation)

This patch has also dropped support for toplevel variables, which will be added
back at some point. I wonder if we should try to also implement full lexical
closures and lambdas when we do that. The current implementation of local vars
is already halfway there, and having proper lexical closures would definitely
allow for a cleaner and less hacky implementation of expressions inside the
toplevel initializers.