Introduce lambda expressions (but not closures)
This introduces an ast.Exp.Lambda type and changes the Toplevel.Func nodes to
use it. This was made to clean up the code a little bit, but should also help
pave the way for full lambdas in the future.

Since this commit did not add lamda expressions to the parser and typechecker,
there should be no visible changes to the user yet.