Implement ToIR pass
This pass converts the typed AST to the sequence of commands.

Originally I considered merging this with the typechecker, which would remove
the need for the _name and _type annotations. However, I changed my mind because
doing it in two passes allows for some useful optimizations:

  - When checking exp in `local x = exp`, we don't know the type of x yet.
    This makes it more difficult to perform that dst_opt optimization we do.

  - When calling a toplevel function or a builtin, we typecheck the "f" but
    don't want to generate any IR for it.

If one day we start optimizing these things in a later optimization pass, we
might reconsider this choice.

It goes without saying that since our coder tests are currently disabled the
test coverage for to_ir is less than ideal. So I expect to have to fix some bugs
once the new coder comes online.