upvalue record -> upvalue userdata array
Revert the upvalue representation to the same one used by the old coder.
The attempt at using records was both opintless and complicated.

It was pointless because the only advantage of using records would be if we
had module-local variables with primitive types, which are not performance
sensitive.

It was complicated because we needed to special case the upvalues record when
generating code. For example it did not have a constructor function.

Additionally, the global upvalues are initialized in the luaopen function using
the vanilla C-API instead of the low-level pallene-generated code. This means
that even if we used records we ended up duplicating a bunch of code anyway.