Move some things to pallene_runtime.h
Let's put all out #include directives in one place isntead of spread out all
over the place. We can also define some division and bitshift as inline
functions instead of as a big chunk of C code. Not only is this more robust
but it also makes the coder.lua and the generated C code a bit easier to read :)