More strict util.render: forbid numbers
Now the replacement values must strings. Integers and floating point numbers are
not allowed.

Previously it was very easy to accidentally pass numbers to util.render, without
wrapping them with C.integer or C.float first. This isn't a big issue with
integers, but is a silent error with floating point numbers.