Return variable is now a regular variable.
Previously we had a separate return variable named "ret" and the return
statement assigned to it. This often resulted in an extra assignment in the
level of the generated C code:

    x1 = blah();
    ret = x1;
    return ret;

Now the return variable is a regular variable like the other ones so there is
no need for the extra assignment.

    x1 = blah();
    return x1;