Fix some clang warnings about flags
The --std flag had a typo was actually supposed to be -std. Fixed that.
The -pedantic flag is also known as -Wpedantic. Changed that to be consistent.

Clang 8 used to give us -Wunused-command-line-argument warnings because we
were passing C-language flags like -std to the assembler and linker. So I
changed c_compiler.lua to only pass the flags that we actually need to each
step. The code ended up a bit cleaner as a bonus. :)