Add ir.Seq
In the previous representation of ir.Cmd, all our recursive procedures needed to
be a pair of mutually recursive functions -- one for ir.Cmd and one for lists of
ir.Cmd

By adding an ir.Cmd.Seq operation we can now turn all of our lists of commands
into a single command, and about the need for mutually-recursive functions.

Since many pieces of code check for empty lists of commands, I also added a
ir.Cmd.Nop command. Yes, it is redundant with ir.Seq but ir.clean takes care
of that.