Implement constant propagation of globals
This commit implements an optimization that removes module variables that are
constant or that are never read from. `GetGlobal` reads from constant variables
removed this way are replaced by a `Move` of the corresponding value.

For example, in the following code the N variable is optimized away, as if we
had written `for i = 1, 10`:

    local N = 10
    function f()
        for i = 1, N do
            blah()
        end
    end

Furthermore, in the following code the `x` variable is not created in the final
version of the code, since it is never read from. The `blah()` function still
gets called inside `f`, but its return value is thrown away.

    local x = 42
    function f()
        x = blah()
    end

The important limitation for now is that a module variable is only identified as
a constant if its initializer is a literal value (number, string, etc). This
means that in the following program `x` and `y` are NOT considered constant.

    local x = 1 + 1
    local x = y