Change how stack frames work, and optimize dead GC
This commit has two major changes. The first one is a change to how stack frames
work and the second one is an optimization for GC variables.

The previous version of Pallene used to set L->top once at the start of a
function, to reserve enough space for all the GC vars we might possibly need.
However, we did not set the slots to `nil` so the GC could potentially see some
garbage values from old function calls.

In the new version, we instead change L->top on demand to reflect the actual
number of GC variables that we have saved on the stack. More specifically, we
now set L->top to its proper value before running the GC and before calling Lua
functions. This is similar to how lvm.c does it.

As a part of this first change, I also modified the type signature for
the Pallene calling convention to add the stack base pointer as a parameter.

Additionally, I also implemented an optimization to avoid saving variables to
the Lua stack if they die before the GC has a chance to run. That said, I think
there is still some room for tweaking how this works. For details, please read
the code comment on pallene/gc.lua

(ps.: it turns out that due to how the stack allocation algorithm works it was
easier to do these two changes in one commit than separately)