Implement uninitialized variables analysis
This commit adds the option to declare uninitialized variables. For example

    local x: integer
    if foo then
        x = 10
    else
        x = 20
    end

As a part of this, it also replaces the old "always_returns" pass from the
checker with a new pass at the IR level that checks for variables that are used
before being initialized, and for functions that reach the end without returning
a value.