sizearray -> alimit
The latest version of Lua has changed the "sizearray" field into an "alimit"
field, to optimize the implementation of the `#` operator.

The basic idea is that alimit is less than or equal to the old sizearray, and
doubles as a good guess for the value of `#`.

The important thing for us is that it is still true that if (i < alimit) we can
access field i in the array part, and that calling resizearray will increase
alimit as expected. So there is no need to change our stategy for accessing
table fields.