Allow nils in arrays
This PR has two changes

The first one is that now we allow arrays of nil, aka {nil}.

I believe that the original motivation for this restriction was because it was
impossible to tell apart in "in-bound" nil from an "out of bound" nil.
However,we have since changed our interpretation of Pallene to not have a
separate "array out of bounds" error and and treat "out of bounds" nils no
differently than an in-bound nil: both are just a run-time tag errors. Under
this interpretation there is no need to forbid {nil}.

Furthermore, even if we wanted to keep the concept of "out of bounds" there
would be no need to forbid {nil} anymore. The latest version of the Lua
interpreter uses a variant LUA_TEMPTY tag to distinguish the regular nils from
the nils that live out of bounds or in array holes.

This brings us to the second change in this patch. Under the previous version of
Pallene, we would return an "empty" value when reading from an out-of-bounds
index. This would result in strange behavior because this "empty" value is not
considered to be == nil !

    function get(xs: {value}, i:integer): value
        return xs[i]
    end

    -- Lua
    local x = test.get({}, 1)
    print(x)        --> nil
    print(nil == x) --> false (!)

This patch corrects this by ensuring that we always clear the LUA_TEMPTY tag
when we read a nil from an array. At first glance it might sound wrong to have
to do a tag check when reading from an `{value}` but rawget also has to do it so
I think it is unavoidable.

Closes #89