Implement string comparison
Closes #91

The nontrivial part of this patch is that l_strcmp, the function in lvm.c
responsible for string comparison, is declared as a static function, and
therefore cannot be called by Pallene.

Our options were to either redeclare l_strcmp as a public function, or copy it
into the Pallene runtime library. I chose the latter alternative, to minimize
our number of custom Lua patches.