Change how we print blank lines inside functions
Now newlines inside functions are removed by default. For the handful of cases
where we actually want a blank line in the output the formatter will now convert
empty /**/ comments into a blank line.

In the process, I changed many of the util.render calls in coder.lua because now
we can write them without worrying if they are leaving a "\n" at the end or not.

(The formatter is not indempotent anymore. A line with /**/ becomes a blank line
in the first pass, and is removed in the second pass).