Fix floating point literals that are also integers
After I changed the C.float() function to print floating point numbers with
string.format("%g") instead of string.format("%a"), numbers like 0.0 and 1.0
started being printed as 0 and 1, without a decimal dot.

Most of the time, this did not make a difference because C automatically coerces
integer literals to double where needed. However, it produced the wrong result
when things like (1.0/2.0) in Pallene were converted to (1/2).

In addition to being more correct, adding a decimal dot to the floating point
literals in C also makes the generated C code a bit easier to read.