Move tag checking inside the get_slot function
Previously, we always had to be sure to check the tag of a value before we used
the get_slot function. It was easy to forget to do this, as happened in the case
of the return value of Lua functions.

This commit refactors the get_stack_slot and get_luatable_slot functions into
three functions:

* unchecked_get_slot
* Checker:get_stack_slot
* Checker:get_luatable_slot

When reading a value that comes from Lua (that is, from the Lua stack or from a
Lua table), the appropriate function now automatically does a tag check. When we
know the value is already the type we want (for example, in record fields) then
we can use the unchecked_get_slot function, whose name makes it very clear that
there is no tag checking.