Check if src_fields and dst_fields are right
In addition to the src_fields and dst_fields list, there is now also an
other_fields list, and some code that checks that all field names are accounted
for.

This should help prevent bugs similar to #166 from appearing again. It has
already helped to find out that we had forgotten about the "src3", which we
fixed in the previous commit.