Make return values be parameters of ir.Cmd.Return
In this PR we change the `ir.Cmd.Return` instruction to receive the return
values as a list of `ir.Value`.

Currently, the return values are implicit. They are assigned to a special group
of local variables, just like function arguments are, and it is implicitly
assumed that the return statement returns whatever is stored in these local
variables.

IIRC, the original motivation for this was because the book that I was referencing
at the time did it that way. One way to detect if the control flow reaches
the end of a non-void function without returning a value is to check if the
"return variable" is uninitialized at the end of the function.

However, this solution has proved to be a bit of a foot-gun because
`ir.get_srcs` for ir.Cmd.Return currently returns the empty list. This had
already caused a bug in the GC (PR #136), requiring special treatment in gc.lua.
Furthermore, today Leonardo ran into the same problem, while implementing
some other static analysis in the code.

To add insult to injury, we weren't even using that trick with the uninitialized
return variable, which was the whole reason why we were originally doing this.
In uninitalized.lua, we were separately keeping track of whether the function
can fall through through to the bottom without reaching a return statement
first.

So let's solve this problem for good! In this patch, the return values are now
passed as parameters to the `ir.Cmd.Return` instruction, so `ir.get_srcs` should
do the right thing.