Added support for multiple returns
Functions with multiple returns get expanded and type checked when
needed. The function expansion occurs before the checks, making them
validate the number in the lhs with the number in the rhs of an assignment
or a declaration. If a function call has the last argument aa function with
multiple returns, it is expanded and then validated for number of arguments
and types.

Dynamic call is a work in progress.