Refactored multiple returns code after review
Unified how the function gets expanded on declarations, assignments,
and toplevel declarations. Changed coder code to match the new
model.

Only the necessary function returns needed to match the lhs are
expanded. Same rules from Lua applies, expansion only occurs if
the function call is the last expression on the rhs or the last
parameter on a function call.