C.reformat now ignores '{'. Also uses LPEG.
C.reformat used to indent the program if it encountered a brace or paren inside
a character literal. This commit tells C.reformat to also ignore the contents of
a character literal.

We also updated the code that ignores comments and literals to use LPEG instead
of a bunch of gsubs. Hopefully it should be a bit more robust now.