Create less temporaries in multiple assignments.
This commit improves the compilation of multiple assignments by using
exp_to_assignment instead of exp_to_value, when posiible. This should create
less temporary variables in many cases.

The current multiple assignment code is smart enough to not create a temporary
variable if the RHS is a variable name that is not reassigned to in the multiple
assignment. It already does a good job in situations such as the following,
where it creates a temporary for the "y" in the RHS but not for the "x".

    x, y = y, x

However, since the old algorithm always used exp_to_value` it inadvertently
created temporaries for anything that fell in the exp_to_assignment case. For
example, the following assignments would create a needless temporary variables:

    x = x + 1
    x, y = f()

The second example is specially annoying because it invalidated our clever
implementation of the ExtraRet case in exp_to_assignment.