Initialize our TValue local variables
Booleans are now similar to Nil in that they only use the type tag part of the
TValue and don't initialize the ".value" field of the tagged usion. This means
that booleans are now vulnarable to those -W-maybe-uninitialized errors that we
had encountered with Nil.

Implementing the workaround that we used for nils in pallene_setbvalue would not
be very appealing because pallene_setbvalue is called much more often than
pallene_setnilvalue, meaning that adding an extra assignment to `.value` would
be wasteful. So I bit the bullet and just initialized all our TValue local
variables now. In theory the C compiler should be able to optimize away these
initializers if it detects that the values are overwritten before they are read
from.