size_hint should be a ir.Value
Currently the size_hint is an integer, which works fine for tables and arrays
whose length is known at compile time but won't work well for tables whose
length is only known at run time, like the following one:

    local xs = {}
    for i = 1, N do
        xs[i] = 0
    end

Having size_hint be an ir.Value will allow us to use the value of `N` as the size
hint for `xs`. This commit doesn't implement that optimization just yet but it
paves the way for it.