Increase source code column limit to 100
A wider limit of 100 reduces the number of times where we need to do special contortions to fit
inside the column limit, while still being narrow enough to fit multiple windows side-by-side and to
keep block comments easy to read.

While working on this commit I noticed that there were a bunch of places where we were already going
over the limit of 80 columns. I propose that we stop fighting with ourselves and just relax the
limit to 100 instead, which is something that we will hit much less often.