Create IR variables for "unused" return values.
Fixes #233.

In the previous implementation, the dsts array for function call instructions could contain `false`
entries, corresponding to return values that were not assigned to any local variables. It was the
job of the coder to take care of that. The problem was that we forgot to correctly handle these
missing destination variables in the case of built-in function calls, meaning that throwing away
the return value of a builtin call would crash the Pallene compiler:

    function oops()
        math.sqrt(4.0)
    end

This commit fixes this crash by getting rid of those `false` entries in the dsts array. The to_ir
step now creates temporary variables for these cases.

This commit also changes Pallene's semantics in the corner case where a CallDyn function call
returns the wrong type for a return value that is not used. We reverse the decision we made in #165
and #167 because now the IR does not make a distinction between used and unused return values.