Workaround busted not gracefully handling SIGINT
Fixes #236.

If the test suite runner is interrupted with SIGINT, the teardown routines of the test suite are not
run and some files such as __test__.pln, __test__.lua, etc are left in the working directory. This
can happen if the developer presses Ctrl-C a bunch of times to cause busted to stop running, perhaps
because they noticed that all the tests are failing because they forgot to run `make` first.

Unfortunately, I don't think there is a way to tell busted to gracefully interrupt the test suite if
it is in --keep-going mode. In this commit we work around the problem in two ways:

1. Recommend `busted --no-keep-going` as the default. This way we only need a single Ctrl-C to stop
the test suite, and it likely interrupts a test case instead of interrupting busted itself.

2. Add the __test__ files to the .gitignore, just in case.