Type errors involving function types show a better error message.
This commit implements a solution for issue #231. The problem is, the
compiler generates error messages such as "expected function but
found function in assignment". Such error messages are too vague for
the programmer to understand.

The newly added function join_type_list(), along with types.tostring(),
recursively work together to generate the function shape, which includes
the return type and parameter list.