Added a test case for catching incompatible function type assignments.
In the test, we try to pass reference to a function that accepts two arguments
and returns a value to another function. However, the parameter type of the
target function and the argument type specified by the caller do not match.
Therefore, the Pallene compiler is expected to report the error.