Initialize closures together with strings and metatables
Instead of initializing the closures in a separate step, initialize them in the same place where we
initialize the other kinds of module-level constants (strings and metatables). This simplifies the
coder and also makes exporting functions work more similarly to exporting of other values.

The small price we pay is that if a function is exported but not used as a first-class value it
will now use an upvalue slot while it previously would not. Thankfully it does not lead to memory
leaks because those function closures would already be kept alive by the exports table. It does
cause an insignificant amount of memory bloat though. That said, if we really want to optimize that
then perhaps the right way to do it would be to implement closures that are safe-for-space instead
of special casing the creation of function values in the luaopen.