Top-level declarations cannot have duplicates.
In groups of function definitions that are defined next to each other, Pallene (unlike Lua) allows all the functions to see each other. We need to ensure that in each group of functions there are never two functions with the same name.

The second issue has to do with exported functions and variables. We need to ensure that those names are not shadowed at the end of the file, which is where we will insert the return statement.