Added a spec file for testing the Pallene to Lua translator.
The Pallene compiler is invoked with the `--emit-lua` flag for the given
test Pallene code. The input source file is named `__test__.pln` and the generated
output is stored in `__test__.lua`. To assert correct translation, the utility
function `assert_translation` can be used. It loads the output file and
compares with the expected translation.