Added the translator module.
If we passed only the Pallene code to the translator, we would have to invoke
the lexer, parser, checker, and optimizer. This is what `driver.compile_internal`
is already doing. The translator requires both the input and the AST to translate
the code. Therefore, `compile_internal` was modified to return the input code
as well.