Updated the translator to handle multiple declarations.
The last_index is now accessible throughout the module. This prevents us from updating it at every
point. It will be moved to an object in a later commit.

Since `add_whitespace()` will most likely be preceded with a call to `add_previous()`, the latter
was updated to invoke the former in this commit.