Added a new test cases to the translator spec.
The .editorconfig file was updated to prevent the translator spec from being trimmed. Otherwise,
a sublte bug is introduced in the spec file. Consider the following example:
```
local xs:
	integer = 10
```
The equivalent translated Lua code would be:
```
local xs
	        = 10
```
Since the colon is replaced with a space in the expected string, leaving a whitespace at the end
of a line, the editor trims it away. This causes test cases with an expected space at the end of
a line to fail.