The translator now uses gsub() to retain tabs, carriage returns, spaces, and newlines.
The general idea is to use `string.sub` to get the substring we want to replace with whitespace.
After which, we call `string.gsub` on the substring with `%S` to replace all non-whitespace
characters with space. Whitespace characters including newlines stay unchanged.

As of now, comments before the actual type annotation begins are retained. See the snippet
below:
```
local i : -- This comment is retained.
{
   -- This comment will be erased.
   integer
} = { 1, 2, 3 }
```
This is possible because we have access to enough offsets to retain such comments.
However, comments that appear within type annotations are erased. We could think
of a workaround later if it turns out that preserving those comments is important
in practice.

The most important thing is preserving newlines because that means error messages
point to the right place. Preserving comments is a nice thing to do but is not
essential because you still have the original Pallene file with all the comments.