The translator now removes return types in function definitions.
To extract the location of the return types starting from the column, we had two options in mind.
The decision was between caring about all the kinds of type or caring about all the kinds of
places where a type can appear in. The argument for the latter was, I would have to rewrite
everything else to translate the return types. Also, the location information is used only by
the translator. So it would not really matter if the locations were not perfectly organized.
Therefore, the latter option was chosen.

To extract the locations of the return types, the `toplevelfunc` rule was modified to capture the
location. This is similar to what we did with the `type` rule, with the `decl` being the parent
rule.

Further, the translator spec was updated to include tests for function return types.