The translator now generates code for exporting symbols.
All the symbols that are exported are added to an auxillary sequence. A return statement at the
end of the module with all the exported symbols is added. Since the statement is added at the
end, the line numbers from the original input are conserved. Further, the return statement is
not generated if there are no exports.