Clarify syntax error for expected ')'
Sometimes we get a syntax error because there is an unexpected token before the ')'. In those cases,
it might not be accurate to complain that the ')' is missing or that the parenthesis are unmatched.

For example, the following function call is parsed as a function name followed by an empty parameter
list followed by an unexpected while token. The ')' is there, but we got a syntax error before we
reached it.

   f(while)